Website/phone-intelligence/requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.38.0
datasets>=2.14.0
scikit-learn>=1.3.0
accelerate>=0.24.0
joblib>=1.4.0
pandas>=2.2.0
requests>=2.31.0

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0

# Phone / message analysis
phonenumbers>=8.13.0
Pillow>=10.0.0
numpy>=1.24.0

# OCR options
pytesseract>=0.3.10
easyocr>=1.7.1
rapidocr-onnxruntime>=1.3.24
opencv-python-headless>=4.8.0.76
torchvision>=0.15.0

# PII redaction
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.6.0

# Utilities
tqdm>=4.66.0

# Tests
pytest>=8.0.0
